Every estimate now maps to its own RealTimeTrip object
Before, I had all the estimates for a leg globbed up in a RealTimeLeg object contained in a RealTimeTrip object. Now, each estimate is mapped to its own RealTimeTrip and its respective RealTimeLeg object. This is so that the user can immediately see all of the estimates at once, instead of having to click around to see them. Doing so will minimize the amount of clicks a user has to preform in order to get the information they need about a future trip.