Fixed bug related to "action bar" title being set improperly and removed extra call to requestTrips in RealTimeTripsFragment
Instead of delegating to all RealTimeTripsFragments maintained by the FragmentManager to set the "action bar" title, now only the last fragment in the stack will set the title. This is because I was encountering some issues with an incorrect title set due to the FragmentManager's list of fragments not matching the current stack of BottomNavigationView's backstack and it doesn't make sense for non-top-of-the-stack fragments to do things because if they're not at the top, they're not visible.