Removed the idea of "incomplete" RealTimeTrip/RealTimeLeg
tl;dr replacing spaghetti related to gluing together different api calls with some client sided math

In order to minimize the amount of API calls and spaghetti glue needed to get a fully fleshed out trip, we'll calculate out a best estimate for a trip. This means that all trips will be "completely" rendered with the initial fetch to them, which means we no longer need the idea of "incomplete" models. This change also lead to the deletion of the `RealTimeTripInfoViewModel`, since it's no longer needed to "complete" an "incomplete" trip.